days/day-24-project-production-pipeline/requirements.txt
# Day 24: Production Pipeline - Dependencies
apache-airflow==2.7.0
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-redis==3.4.0
dbt-core==1.6.0
dbt-postgres==1.6.0
great-expectations==0.17.23
boto3==1.34.0
pandas==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.7
redis==4.6.0
prometheus-client==0.17.1
python-dotenv==1.0.0
pytest==7.4.0
faker==19.6.2
localstack-client==2.5
awscli-local==0.22